RAG before interim report
This code is mainly used to extract relevant information from large amounts of text data to answer a series of questions. First, it uses regular expressions and sorting logic to split the raw data into smaller chunks of text. Then, a SentenceTransformer is utilized to generate embedding representations of these text blocks, and the faiss library is used to create an index for fast similarity searches. Next, the program retrieves the text blocks most relevant to the query given the query and the created knowledge base. Finally, these relevant chunks of text are used as context for the OpenAI GPT-3.5 model to generate answers to a series of questions. The code also includes the necessary library imports and configurations, as well as the definition of some functions for processing text blocks, creating knowledge bases, retrieving related text blocks, and generating answers using the GPT-3.5 model